+++ test/unit/state/move-cross-axis/get-closest-draggable.spec.js
