+++ test/unit/state/move-cross-axis/move-to-new-droppable.spec.js
@@ -1,0 +1,5 @@
+// @flow
+
+describe('move to new droppable', () => {
+
+});
